book-service/requirements.txt
Django==4.2.27
djangorestframework==3.15.0
